+++ lib/widgets/dialogs/settings_dialog.dart
@@ -22,6 +22,8 @@ import '../../services/file_service.dart';
 import '../../services/recovery_service.dart';
 import '../../services/classification_enforcement_policy.dart';
 import '../../services/git/git_forge.dart';
+import '../../utils/net_guard.dart';
+import 'certificate_trust_dialog.dart';
 import '../../services/s3/s3_service.dart';
 import '../../services/webdav_service.dart';
 import '../../models/local_cve_status.dart';